Update getting-started.md (#446)
A slight addition to step 6 that helps explain how to connect up the category properly. This information would only really be helpful to a total beginner who doesn't know how json works... And I'm exactly that, and I know I'm not the only one. (I conferred with an expert on this, I can confirm the information I added is accurate and works)